szkol_referencja/copilot_training_self_paced/10_copilot_python_sdk/solutions/requirements.txt
github-copilot-sdk>=0.3.0
pydantic>=2.5
pytest>=8
pytest-asyncio>=0.23
